Made a lot of progress

+++ Dobri Tasklist Manager/Dobri Tasklist Manager/Controllers/TaskListsController.cs
@@ -22,7 +22,9 @@ namespace Dobri_Tasklist_Manager.Controllers
         // GET: TaskLists
         public async Task<IActionResult> Index()
         {
-            return View(await _context.TaskLists.ToListAsync());
+            User currentUser = _context.Users.FirstOrDefault(x => x.Id == Active.CurrentUserId);
+            var currentUserTaskLists = _context.TaskLists.Where(x => currentUser.ListsCreatedByMe.Contains(Convert.ToString(x.Id)) || currentUser.ListsSharedWithMe.Contains(Convert.ToString(x.Id)));
+            return View(await currentUserTaskLists.ToListAsync());
         }
 
         // GET: TaskLists/Details/5
@@ -43,6 +45,23 @@ namespace Dobri_Tasklist_Manager.Controllers
             return View(taskList);
         }
 
+        public IActionResult Share(int? id)
+        {
+            Active.CurrentTaskListId = (int)id;
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Share(int id, User user)
+        {
+            user = _context.Users.FirstOrDefault(x => x.Username == user.Username);
+            user.ListsSharedWithMe += "," + Convert.ToString(id);
+            _context.Users.Update(user);
+            await _context.SaveChangesAsync();
+            return RedirectToAction("Index");
+        }
+
         // GET: TaskLists/Create
         public IActionResult Create()
         {
@@ -56,6 +75,12 @@ namespace Dobri_Tasklist_Manager.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,DateOfCreation,DateOfLastEdit,IdCreator,IdLastEditor")] TaskList taskList)
         {
+            User currentUser = _context.Users.FirstOrDefault(x => x.Id == Active.CurrentUserId);
+            currentUser.ListsCreatedByMe += "," + Convert.ToString(taskList.Id);
+            taskList.DateOfCreation = DateTime.Now;
+            taskList.DateOfLastEdit = DateTime.Now;
+            taskList.IdLastEditor = Active.CurrentUserId;
+            taskList.IdCreator = Active.CurrentUserId;
             if (ModelState.IsValid)
             {
                 _context.Add(taskList);
@@ -63,6 +88,12 @@ namespace Dobri_Tasklist_Manager.Controllers
                 return RedirectToAction(nameof(Index));
             }
             return View(taskList);
+        }
+
+        public IActionResult Tasks(int? id)
+        {
+            Active.CurrentTaskListId = (int)id;
+            return RedirectToAction("Index", "ToDo");
         }
 
         // GET: TaskLists/Edit/5
@@ -88,6 +119,8 @@ namespace Dobri_Tasklist_Manager.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name,DateOfCreation,DateOfLastEdit,IdCreator,IdLastEditor")] TaskList taskList)
         {
+            taskList.DateOfLastEdit = DateTime.Now;
+            taskList.IdLastEditor = Active.CurrentUserId;
             if (id != taskList.Id)
             {
                 return NotFound();
@@ -140,8 +173,34 @@ namespace Dobri_Tasklist_Manager.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var taskList = await _context.TaskLists.FindAsync(id);
-            _context.TaskLists.Remove(taskList);
-            await _context.SaveChangesAsync();
+            User currentUser = _context.Users.FirstOrDefault(x => x.Id == Active.CurrentUserId);
+            if(taskList.IdCreator != currentUser.Id)
+            {
+                string sharedLists = "";
+                foreach (var item in currentUser.ListsSharedWithMe.Split(','))
+                {
+                    if(item != Convert.ToString(taskList.Id))
+                    {
+                        if(sharedLists == "")
+                        {
+                            sharedLists = item;
+                        }
+                        else
+                        {
+                            sharedLists += "," + item;
+                        }
+                    }
+                }
+                currentUser.ListsSharedWithMe = sharedLists;
+                _context.Users.Update(currentUser);
+                await _context.SaveChangesAsync();
+            }
+            else
+            {
+                _context.TaskLists.Remove(taskList);
+                await _context.SaveChangesAsync();
+            }
+
             return RedirectToAction(nameof(Index));
         }
